refactor(renderer): isolate official favorite dialog entrypoint [skip ci]
