+++ desktop/renderer/index.html
@@ -731,7 +731,6 @@
   <script src="routing-manager.js"></script>
   <script src="certificate-manager.js"></script>
   <script src="group-dialog.js"></script>
-  <script src="official-favorite-dialog.js"></script>
   <script src="add-website-dialog.js"></script>
   <script src="campus-service-workspace.js"></script>
   <script src="proxy-auth-migration.js"></script>
